simple test case (Doesn't work)
Here is a simple test case to ensure we run the same code. The binary doesn't execute for me on mac. Maybe you can convert the text files, then reupload and add a checker to the shell script to only run the convert if the text files are not present.